유저가 자기 자신을 팔로우 할 수 없도록 수정

+++ src/main/java/org/example/newsfeedteamproject/user/service/FollowService.java
@@ -30,6 +30,11 @@ public class FollowService {
         // 엔티티 자체를 가져오기 떄문에 Optional findBy...() + Optional<Follow> 사용
         Optional<Follow> existingFollow = followRepository.findByFromUserAndToUser(fromUser, toUser);
 
+        if (toUserId.equals(fromUser.getId()) ) {
+            throw new IllegalArgumentException("자기 자신은 팔로우 할 수 없습니다.");
+            // IllegalArgumentException -> 잘못된 인자 접근시 사용되는 런타임 예외
+        }
+
         if (existingFollow.isPresent()) {
             // 팔로우 되어 있으면 언팔로우 처리
             followRepository.delete(existingFollow.get());
